Hack to fool your friends
Copy the WHOLE script and make a new bookmark. Past the code on in the url box and once you are in the real game (not the lobby), your list of hacks will show up. Have fun hacking. I will sponsor Glixzzy for this amazing hack he shared with us.